requirements.txt
asgiref==3.5.2
Django==3.2.16
django-currentuser==0.5.3
django-debug-toolbar==3.7.0
Pillow==9.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
